Data Id & Class Formatting
Fixed the html_lib_use(); function as it ignored id's on elements that were not a div eg: h1, p, img tags. Updated this function to also create classes from a data attribute. So now when this html data is used it should remain unstyled by the css. Updated the game_lib.html to be compatible with this build